Make highscore character limit configurable (4-9 characters) in Simply Love Options
This commit allows machine operators to select the limit to the number of highscore characters that are selectable for players. The default is 9, but the option to select any number between 4 and 9 is available. I strongly believe that if players are going to use offensive words for their highscore name, they already would be doing so. It only takes 4 letters to type "FUCK", and even if you set the limit to 3, "FCK" is still "offensive". 9 characters should give players a lot more leeway to use more fitting nicknames, and even fit most of first names into a highscore name.

9 is the most characters that would fit comfortably, but with a bit of reclaiming dead space in the theme I can't imagine 20-25 characters being much of a problem (yay full names?)