New "Has Edit" graphic to give more room for song titles
It's really bothered me that room is made for a "Has Edit" image on the music wheel taking away from the space available to Titles/Subtitles. This commit makes a new image and places it beside the music wheel so that more room is left over for titles so that long titles don't overflow onto this image.

In the future, I'd like to completely remove this icon because it is a very poor workaround to not being able to see ALL of the charts for a particular song. In the future, it would be best to move to an interface system that more closely resembles the default Stepmania theme whereby you can always see the edit charts without having to check with ⬇️ ⬇️ . This however would have to call a lot of UI elements into question, and for the time being, this solves enough of the problems.
 Home