Create ScreenSelectMusicWide
This will act as a base for the ScreenSelectMusic rework. I modified SL-Branches to forward a new derivative of ScreenSelectMusic, "ScreenSelectMusicWide" when in a widescreen aspect ratio. This will lay the groundwork for a rework of ScreenSelectMusic that takes more advantage of the space offered by WideScreen. I imagine that in the future, if ultrawides become more popular, a "ScreenSelectMusicWide_XL" could provide even more utility with the huge amount of extra screen real-estate.

As far as I can tell, there are no issues with Branching, and everything seems to be working as expected from ScreenSelectMusic.