Add eBPM readout to SSM(Wide) and Gameplay
Add and eBPM (effective BPM) readout on ScreenSelectMusic and ScreenGameplay, located to the right of the Peak NPS readout. eBPM is equal to the peak NPS in a chart, whereby said NPS is represented by a BPM where 16th notes equaled the same NPS. Peak NPS is a relatively new metric that tells the player what quantization of notes they can expect to encounter at the peak of a chart's difficulty, but is a "fun number" that non-players can easily understand; Peak NPS is far more abstract to skilled Stepmania players. It is much less abstract to the skilled Stepmania player to consider the eBPM and ascertain as to what quantization a chart follows; a 110bpm chart with a max eBPM of 165 tells the player that there are 24th note sections. eBPM also helps in the case of "skittles" charts to tell the player what BPM they will be playing at; many skittles patterns are either very difficult to visually discern because of the lack of quantization colours available in Stepmania (20th notes are a particular painpoint, at this time).

Math:
-If eBPM = BPM/2, then the chart's hardest section contains 8th notes
-If eBPM = BPM, then the chart's hardest section contains 16th notes
-If eBPM = BPM * 1.5, then the chart's hardest section contains 24th notes
-If eBPM = BPM *2, then the chart's hardest section contains 32nd notes.

Disclaimer: Peak eBPM will be incorrect if sections of peak NPS are too short. This was not an issue for NPS readouts. In the future it may be worth looking into parsing a chart based on the quantization, rather than just purely how many notes appear per second.

Also: the maxwidth on ScreenSelectMusicWide was too snug to the edges of the player's song info card. This gives a small amount of padding when a chart has a lot of chart information to display.