Show a preview of the selected chart on SSMw
This feature requires OutFox 0.4.18 LTS or newer.

Def.Notefield will respect most mods selected by the player (mods that add/remove steps won’t display properly in the preview). This feature is fantastic for players of all skill levels to get a more accurate representation of the chart being hovered. The portion of the chart that will be displayed is dependent upon the chart’s “#SAMPLELENGTH” and “#SAMPLESTART” tags.

Much of the UI had to be shifted around in 2 player mode to make the previews fit onscreen.

Below are a number of notes from the pre-quashed commits:

If a player uses any Mini% it throws off the positioning on the notefield. By forcing the notefield to display 0% mini, I don't have to compensate for a player's Mini% to make the notefield display within its boundaries.

"Since you're not using the player argument in this lua file, you can safely take the LoadActor away from the loop."

Joining a new player to ScreenSelectMusicWide is going to be janky because the best way to go about it would be to fade out the first joined player's UI. We'll have to then deal with the performance hit associated with having the UI be duplicated but not visible (because hidden elements are still loaded?). The best option, I think, is to reload the screen entirely.

Use a local function to determine the X/Y position and zoom of the Chart preview based on the UI elements onscreen.

Co-Authored-By: Jose Varela <josevarela@tutanota.com>
Co-Authored-By: Sudospective <avery@sudospective.net>
Co-Authored-By: MrThatKid <18648486+mrthatkid@users.noreply.github.com>