What does this module do? Getting rid of it...
I don't know what this module does; it chokes when trying to load certain player profiles getting a nil value. Its purpose was supposed to load the speed mod for each player and account for 0BPM DISPLAYBPM charts? It looks like Def.Notefield takes care of this itself in the same way that mods (Mini, noteskins, etc.) are loaded properly.