StepmaniaX (smx) support
This commit is a modification of the pull request here: https://github.com/Horsey-/Horseys-Simply-Love/pull/12

This commit adds StepmaniaX support. You might still need to download StepmaniaX specific noteskins (there are some that come bundled with OutFox, but are not at all similar to any of the official noteskins as far as I know)

Beat bars are hard-enabled in metrics by this commit. In the future I'd like to add support for beat bars in all game modes and on a per player preferential basis.

Timing is all based on the ITG FA+ (ITL) standard. This could be amended in the future; I'm not aware of any differences in judgement timing.

Co-Authored-By: Jose Varela <jose@josevarela.net>
Co-Authored-By: Tyler Brekke <concubidated@users.noreply.github.com>
Co-Authored-By: JustMoneko <63615036+JustMoneko@users.noreply.github.com>