WIP - Fix ScreenEvaluation
These changes make the theme usable, but there are still errors that need to be ironed out that relate to ITGm specific engine APIs that are not being found by the theme. The issues stem from the calculation of steps with "the boys" which is an engine API in ITGm that doesn't exist in OutFox.

Co-Authored-By: Jose Varela <jose@josevarela.net>